Move package
and remove JsonValue because not needed.
