+++ src/test/kotlin/com/fasterxml/jackson/module/kotlin/_integration/deser/value_class/deserializer/has_json_value/SpecifiedForObjectMapperTest.kt
@@ -2,6 +2,9 @@ package com.fasterxml.jackson.module.kotlin._integration.deser.value_class.deser
 
 import com.fasterxml.jackson.databind.module.SimpleModule
 import com.fasterxml.jackson.module.kotlin.MissingKotlinParameterException
+import com.fasterxml.jackson.module.kotlin._integration.deser.value_class.NonNullObject
+import com.fasterxml.jackson.module.kotlin._integration.deser.value_class.NullableObject
+import com.fasterxml.jackson.module.kotlin._integration.deser.value_class.Primitive
 import com.fasterxml.jackson.module.kotlin.jacksonObjectMapper
 import com.fasterxml.jackson.module.kotlin.readValue
 import org.junit.jupiter.api.Assertions.assertEquals
